Moving big categorical tests to tests_slow folder
