+++ .github/workflows/main.yml
@@ -703,6 +703,7 @@ jobs:
           mkdir -p build/coverage
           (cd build/coverage && cmake -E env CXXFLAGS="--coverage" cmake -DBUILD_PYTHON=1 -DBUILD_PARQUET_EXTENSION=1 -DENABLE_SANITIZER=0 -DCMAKE_BUILD_TYPE=Debug ../.. && make)
           (cd tools/pythonpkg/tests && python3 -m pytest)
+          (cd tools/pythonpkg/tests_slow && python3 -m pytest)
           build/coverage/test/unittest
           build/coverage/test/unittest "[intraquery]"
           build/coverage/test/unittest "[interquery]"
