+++ tools/pythonpkg/tests/pandas/test_pandas_category.py
@@ -84,18 +84,6 @@ class TestCategory(object):
             category.append(str(i))
         check_create_table(category)
 
-    def test_category_string_uint16(self, duckdb_cursor):
-        category = []
-        for i in range (300):
-            category.append(str(i))
-        check_create_table(category)
-
-    def test_category_string_uint32(self, duckdb_cursor):
-        category = []
-        for i in range (70000):
-            category.append(str(i))
-        check_create_table(category)
-
     def test_category_fetch_df_chunk(self, duckdb_cursor):
         con = duckdb.connect()
         categories = ['foo','bla',None,'zoo', 'foo', 'foo',None, 'bla']
